Revert "Bringing gas costs to home page"
